Extract "last message" logging into own producer
The `LAST_MESSAGE` table will be useful for more than just substitution,
and this change makes it more easier to reason about and maintain the
table as separate from the substitution producer.
