+++ src/main/scala/sectery/Producer.scala
@@ -81,7 +81,8 @@ object Producer:
         ),
         Btc,
         Version,
-        Tell
+        Tell,
+        LastMessage
       )
     )
